0.70 final alpha
--HG--
extra : convert_revision : svn%3Ac3e66f6f-232e-0410-82ea-bf51f2194cab/trunk%4077

+++ nggextractXML.php
@@ -20,7 +20,7 @@ global $wpdb;
 $ngg_options = get_option('ngg_options');
 
 // get the gallery id
-$galleryID = attribute_escape($_GET['gid']);
+$galleryID = (int) attribute_escape($_GET['gid']);
 
 // get the pictures
 if ($galleryID == 0) {
